requirements.txt
aws-cdk-lib>=2.0.0
constructs>10.0.0
pytest
black
boto3
pytest-cov